requirements.txt
pandas==2.2.3
streamlit==1.41.1
prophet==1.1.6
cmdstanpy==1.2.4
plotly==5.24.1
folium==0.19.4
streamlit-folium==0.24.0
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
fpdf2==2.8.7
